refactor: 사다리의 높이를 클래스로 포장
