+++ src/main/java/domain/Ladder.java
@@ -7,28 +7,22 @@ import java.util.List;
 
 public class Ladder {
 
-    private final int height;
+    private final Height height;
     private final List<Line> lines;
 
     public Ladder(int height) {
-        validate(height);
-        this.height = height;
+        this.height = new Height(height);
         this.lines = new ArrayList<>();
     }
 
     public void makeLines(int width) {
-        for (int i = 0; i < height; i++) {
+        for (int i = 0; i < height.getHeight(); i++) {
             Line line = new Line();
             line.makeLeg(width);
             lines.add(line);
         }
     }
 
-    private void validate(int height) {
-        if (height <= 0) {
-            throw new IllegalArgumentException(INVALID_LADDER_HEIGHT_EXCEPTION.getMessage());
-        }
-    }
     public List<Line> getLines() {
         return lines;
     }
